Adds feature extractor
This feature extraction workflow runs on videos and extracts a 1x2048
vector of features for each frame in the video. It also uploads the
extracted features to s3 as an hdf file and adds the bucket and object
key as an attribute of the media object for later retrieval and
analysis. Currently, the feature extracted is a placeholder for testing
and will be filled in later with a more meaningful calculation.